+++ CHANGELOG.md
@@ -13,6 +13,11 @@ This is a feature release.
  * Allow fetching all partition offsets for a consumer group by passing a
    `nil` slice in `AdminClient.ListConsumerGroupOffsets`, when earlier it
    was not processing that correctly (#985, @alexandredantas).
+ * Deprecate m.LeaderEpoch in favor of m.TopicPartition.LeaderEpoch (#1012).
+
+confluent-kafka-go is based on librdkafka v2.2.0, see the
+[librdkafka v2.2.0 release notes](https://github.com/confluentinc/librdkafka/releases/tag/v2.2.0-RC1)
+for a complete list of changes, enhancements, fixes and upgrade considerations.
 
 
 ## v2.1.1