+++ CHANGELOG.md
@@ -12,6 +12,10 @@ This is a feature release.
  * Fixes incorrect protofbuf FileDescriptor references (#989, @Mrmann87).
  * [KIP-339](https://cwiki.apache.org/confluence/display/KAFKA/KIP-339%3A+Create+a+new+IncrementalAlterConfigs+API)
    IncrementalAlterConfigs API (#945).
+ * Allow fetching all partition offsets for a consumer group by passing a
+   `nil` slice in `AdminClient.ListConsumerGroupOffsets`, when earlier it
+   was not processing that correctly (#985, @alexandredantas).
+ * Deprecate m.LeaderEpoch in favor of m.TopicPartition.LeaderEpoch (#1012).
 
 confluent-kafka-go is based on librdkafka v2.2.0, see the
 [librdkafka v2.2.0 release notes](https://github.com/confluentinc/librdkafka/releases/tag/v2.2.0-RC1)